Файл разбит по модулям на отдельные файлы
